Fixed a bug in (Function) and (Named Function).
If you existentially quantify a backpointer set in both pre- and
post-conditions, then that set is free to change to anything in the
universe while the command is executing. The corresponding English
statement would be something like "There was a backpointer set before,
and there is a backpointer set now." It says nothing about the
relationship between the two.

What you actually want is a common un-quantified variable in the pre-
and post-conditions. This is then implicitly universally quantified by
the definition of Hoare triples, with a scope that reaches across the
whole triple.

Also, since backpointer sets (like heap cells) have a footprint, you
need to be careful about how they join to the rest of the
formula. Look at the satisfaction relation definition for "forall"
formulae. It requires that the same heap satisfy all the different
formulae. This makes "forall" a bit like iterated \land. What you want
is iterated \sep (which I've written as \bigsep) which means "apply
\sep this many times". Notice that "forall" inside a formula is
defined by your satisfaction relation, while "forall" outside a
formula (for example, in (Object)) is regular mathematics.

To make this work, I used the list operations "length" and
"item". They're easy to define. We can do that later. I also used the
previously mentioned \bigsep. There's a definition of that in the
techreport version of the popl12 paper. It's a theoretically
interesting construct that might be worth chatting about at some
point.
Fixed a bug in (Function) and (Named Function).

If you existentially quantify a backpointer set in both pre- and
post-conditions, then that set is free to change to anything in the
universe while the command is executing. The corresponding English
statement would be something like "There was a backpointer set before,
and there is a backpointer set now." It says nothing about the
relationship between the two.

What you actually want is a common un-quantified variable in the pre-
and post-conditions. This is then implicitly universally quantified by
the definition of Hoare triples, with a scope that reaches across the
whole triple.

Also, since backpointer sets (like heap cells) have a footprint, you
need to be careful about how they join to the rest of the
formula. Look at the satisfaction relation definition for "forall"
formulae. It requires that the same heap satisfy all the different
formulae. This makes "forall" a bit like iterated \land. What you want
is iterated \sep (which I've written as \bigsep) which means "apply
\sep this many times".

To make this work, I used the list operations "length" and
"item". They're easy to define. We can do that later. I also used the
previously mentioned \bigsep. There's a definition of that in the
techreport version of the popl12 paper. It's a theoretically
interesting construct that might be worth chatting about at some
point.