I like this (Object) rule.
I like that you've spotted that the existential when we create new
objects makes it impossible to expand the \bp set before hand using
consequence. I think you're right, and we have to update the \bp sets
in the rules, like this.

In this rule, this has the consequence that the current version of the
rule for:

   {x1:e1,...,xn:en}

..says that none of the expressions e1,...,en create any objects that
have @proto pointers pointing to \lop. This bears thinking about, so
I'm not just correcting it right away - but I suspect using \lop_i to
allow it to change from expression to expression might be all we need
to do to fix it.